# webapp/src/app/model/accessControl.ts
/**
 * Mqtts API
 * Mqtts service
 *
 * The version of the OpenAPI document: 1.0.0
 */

export class AccessControl {
    /**
     * Indicate if we  allow Anonymous access.
     */
    anonymous?: boolean;
}