# webapp/src/app/model/security.ts
/**
 * Mqtts API
 * Mqtts service
 *
 * The version of the OpenAPI document: 1.0.0
 */



export class Security {
    /**
     * Indicate if SSl is Enabled or not
     */
    SSL?: boolean;
    /**
     * Indicates if Client verification is enabled or not
     */
    clientVerification?: boolean;
}